requirements.txt
aiogram>=3.10
reques>=2.31
python-dotenv>=1.0